fix windows ttl bug

+++ src/include.h
@@ -44,7 +44,7 @@ static inline double millitime(){
 static inline int64_t time_ms(){
 	struct timeval now;
 	gettimeofday(&now, NULL);
-	return now.tv_sec * 1000 + now.tv_usec/1000;
+	return (int64_t)now.tv_sec * 1000 + (int64_t)now.tv_usec/1000;
 }
 
 #endif
